typescript(vx-demo): rewrite Streamgraph example in TypeScript
