+++ packages/vx-demo/package.json
@@ -16,6 +16,7 @@
   "author": "@hshoff",
   "license": "MIT",
   "dependencies": {
+    "@types/d3-array": "^2.0.0",
     "@types/nprogress": "^0.2.0",
     "@types/prismjs": "^1.16.0",
     "@vx/annotation": "0.0.192",
